+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jfr/JfrEvent.java
@@ -63,6 +63,8 @@ public final class JfrEvent {
     public static final JfrEvent ObjectAllocationInNewTLAB = create("jdk.ObjectAllocationInNewTLAB", false);
     public static final JfrEvent GCHeapSummary = create("jdk.GCHeapSummary", false);
     public static final JfrEvent ThreadAllocationStatistics = create("jdk.ThreadAllocationStatistics", false);
+    public static final JfrEvent SystemGC = create("jdk.SystemGC", true);
+    public static final JfrEvent AllocationRequiringGC = create("jdk.AllocationRequiringGC", false);
 
     private final long id;
     private final String name;