+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jfr/events/SystemGCEvent.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2022, 2022, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2022, 2022, Red Hat Inc. All rights reserved.
+ * Copyright (c) 2023, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, 2023, Red Hat Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,7 +26,7 @@
 
 package com.oracle.svm.core.jfr.events;
 
-import org.graalvm.compiler.word.Word;
+import org.graalvm.nativeimage.StackValue;
 
 import com.oracle.svm.core.Uninterruptible;
 import com.oracle.svm.core.jfr.HasJfrSupport;
@@ -37,30 +37,26 @@ import com.oracle.svm.core.jfr.JfrNativeEventWriterDataAccess;
 import com.oracle.svm.core.jfr.JfrTicks;
 import com.oracle.svm.core.jfr.SubstrateJVM;
 
-public class JavaMonitorWaitEvent {
-    public static void emit(long startTicks, Object obj, long notifier, long timeout, boolean timedOut) {
+public class SystemGCEvent {
+    public static void emit(long startTicks, boolean invokedConcurrent) {
         if (HasJfrSupport.get()) {
-            emit0(startTicks, obj, notifier, timeout, timedOut);
+            emit0(startTicks, invokedConcurrent);
         }
     }
 
     @Uninterruptible(reason = "Accesses a JFR buffer.")
-    private static void emit0(long startTicks, Object obj, long notifier, long timeout, boolean timedOut) {
+    private static void emit0(long startTicks, boolean invokedConcurrent) {
         long duration = JfrTicks.duration(startTicks);
-        if (JfrEvent.JavaMonitorWait.shouldEmit(duration)) {
-            JfrNativeEventWriterData data = org.graalvm.nativeimage.StackValue.get(JfrNativeEventWriterData.class);
+        if (JfrEvent.SystemGC.shouldEmit(duration)) {
+            JfrNativeEventWriterData data = StackValue.get(JfrNativeEventWriterData.class);
             JfrNativeEventWriterDataAccess.initializeThreadLocalNativeBuffer(data);
 
-            JfrNativeEventWriter.beginSmallEvent(data, JfrEvent.JavaMonitorWait);
+            JfrNativeEventWriter.beginSmallEvent(data, JfrEvent.SystemGC);
             JfrNativeEventWriter.putLong(data, startTicks);
             JfrNativeEventWriter.putLong(data, duration);
             JfrNativeEventWriter.putEventThread(data);
-            JfrNativeEventWriter.putLong(data, SubstrateJVM.get().getStackTraceId(JfrEvent.JavaMonitorWait, 0));
-            JfrNativeEventWriter.putClass(data, obj.getClass());
-            JfrNativeEventWriter.putLong(data, notifier);
-            JfrNativeEventWriter.putLong(data, timeout);
-            JfrNativeEventWriter.putBoolean(data, timedOut);
-            JfrNativeEventWriter.putLong(data, Word.objectToUntrackedPointer(obj).rawValue());
+            JfrNativeEventWriter.putLong(data, SubstrateJVM.get().getStackTraceId(JfrEvent.SystemGC, 0));
+            JfrNativeEventWriter.putBoolean(data, invokedConcurrent);
             JfrNativeEventWriter.endSmallEvent(data);
         }
     }